+++ chain/network/src/lib.rs
@@ -1,5 +1,5 @@
-pub use peer_actor::{EPOCH_SYNC_PEER_TIMEOUT_MS, EPOCH_SYNC_REQUEST_TIMEOUT_MS};
-pub use peer_manager_actor::PeerManagerActor;
+pub use peer::peer_actor::{EPOCH_SYNC_PEER_TIMEOUT_MS, EPOCH_SYNC_REQUEST_TIMEOUT_MS};
+pub use peer_manager::peer_manager_actor::PeerManagerActor;
 pub use routing::routing_table_actor::{
     RoutingTableActor, RoutingTableMessages, RoutingTableMessagesResponse,
 };
@@ -8,9 +8,8 @@ pub use types::{
     NetworkRequests, NetworkResponses, PeerInfo, PeerManagerAdapter,
 };
 
-mod peer_actor;
-mod peer_manager_actor;
-pub mod peer_store;
+mod peer;
+pub mod peer_manager;
 pub mod routing;
 pub mod stats;
 pub mod test_utils;
